chore(deps): update module gitlab.com/gitlab-org/api/client-go to v1

See merge request https://gitlab.com/gitlab-org/cli/-/merge_requests/2600

Merged-by: Timo Furrer <tfurrer@gitlab.com>
Approved-by: Timo Furrer <tfurrer@gitlab.com>

+++ internal/api/issue.go
@@ -7,7 +7,7 @@ import (
 
 // UpdateIssue updates an issue
 // Attention: this is a global variable and may be overridden in tests.
-var UpdateIssue = func(client *gitlab.Client, projectID any, issueID int, opts *gitlab.UpdateIssueOptions) (*gitlab.Issue, error) {
+var UpdateIssue = func(client *gitlab.Client, projectID any, issueID int64, opts *gitlab.UpdateIssueOptions) (*gitlab.Issue, error) {
 	issue, _, err := client.Issues.UpdateIssue(projectID, issueID, opts)
 	if err != nil {
 		return nil, err
@@ -18,7 +18,7 @@ var UpdateIssue = func(client *gitlab.Client, projectID any, issueID int, opts *
 
 // GetIssue returns an issue
 // Attention: this is a global variable and may be overridden in tests.
-var GetIssue = func(client *gitlab.Client, projectID any, issueID int) (*gitlab.Issue, error) {
+var GetIssue = func(client *gitlab.Client, projectID any, issueID int64) (*gitlab.Issue, error) {
 	issue, _, err := client.Issues.GetIssue(projectID, issueID)
 	if err != nil {
 		return nil, err
