+++ internal/commands/changelog/generate/changelog_generate_test.go
@@ -22,9 +22,9 @@ func TestChangelogGenerate(t *testing.T) {
 	gomock.InOrder(
 		tc.MockProjects.EXPECT().
 			GetProject("OWNER/REPO", gomock.Any()).
-			Return(&gitlab.Project{ID: 37777023}, nil, nil),
+			Return(&gitlab.Project{ID: int64(37777023)}, nil, nil),
 		tc.MockRepositories.EXPECT().
-			GenerateChangelogData(37777023, gitlab.GenerateChangelogDataOptions{Version: gitlab.Ptr("1.0.0")}).
+			GenerateChangelogData(int64(37777023), gitlab.GenerateChangelogDataOptions{Version: gitlab.Ptr("1.0.0")}).
 			Return(&gitlab.ChangelogData{
 				Notes: "## 1.0.0 (2023-04-02)\n\n### FirstName LastName firstname@lastname.com (1 changes)\n\n- [initial commit](gitlab-org/cli@somehash ([merge request](gitlab-org/cli!1))\n",
 			}, nil, nil),
@@ -70,9 +70,9 @@ func TestChangelogGenerateWithError(t *testing.T) {
 			gomock.InOrder(
 				tc.MockProjects.EXPECT().
 					GetProject("OWNER/REPO", gomock.Any()).
-					Return(&gitlab.Project{ID: 37777023}, nil, nil),
+					Return(&gitlab.Project{ID: int64(37777023)}, nil, nil),
 				tc.MockRepositories.EXPECT().
-					GenerateChangelogData(37777023, gitlab.GenerateChangelogDataOptions{Version: gitlab.Ptr("1.0.0")}).
+					GenerateChangelogData(int64(37777023), gitlab.GenerateChangelogDataOptions{Version: gitlab.Ptr("1.0.0")}).
 					Return(nil, nil, errors.New(v.errorMsg)),
 			)
