+++ internal/commands/ci/cancel/job/job.go
@@ -90,7 +90,7 @@ func runCancelation(
 			if err != nil {
 				return err
 			}
-			_, _, err = apiClient.Jobs.CancelJob(pid.ID, id)
+			_, _, err = apiClient.Jobs.CancelJob(pid.ID, int64(id))
 			if err != nil {
 				return err
 			}
