+++ internal/commands/gpg-key/get/get.go
@@ -17,7 +17,7 @@ type options struct {
 	gitlabClient func() (*gitlab.Client, error)
 	io           *iostreams.IOStreams
 
-	keyID int
+	keyID int64
 }
 
 func NewCmdGet(f cmdutils.Factory) *cobra.Command {
@@ -50,7 +50,7 @@ func NewCmdGet(f cmdutils.Factory) *cobra.Command {
 
 func (o *options) complete(args []string) error {
 	if len(args) == 1 {
-		o.keyID = utils.StringToInt(args[0])
+		o.keyID = int64(utils.StringToInt(args[0]))
 	}
 
 	return nil
